// admin/src/generated/model/sessionListItemOut.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Embed Chat API
 * OpenAPI spec version: 1.0.0
 */

export interface SessionListItemOut {
  active: boolean;
  created_at: string;
  escalation_status?: string;
  is_hitl: boolean;
  last_activity: string;
  session_id: string;
  visitor_id: string;
}